+++ mockup/js/ui/views/buttongroup.js
@@ -8,6 +8,7 @@ define([
   var ButtonGroup = ContainerView.extend({
     tagName: 'div',
     className: 'btn-group',
+    idPrefix: 'btngroup-',
     disable: function() {
       _.each(this.items, function(button) {
         button.trigger('disable');