+++ mockup/patterns/structure/js/views/actionmenu.js
@@ -3,17 +3,19 @@ define([
   'underscore',
   'backbone',
   'mockup-ui-url/views/base',
+  'mockup-patterns-structure-url/js/models/result',
   'mockup-utils',
   'text!mockup-patterns-structure-url/templates/actionmenu.xml',
   'translate',
   'bootstrap-dropdown'
-], function($, _, Backbone, BaseView, utils, ActionMenuTemplate, _t) {
+], function($, _, Backbone, BaseView, Result, utils, ActionMenuTemplate, _t) {
   'use strict';
 
   var ActionMenu = BaseView.extend({
     className: 'btn-group actionmenu',
     template: _.template(ActionMenuTemplate),
     events: {
+      'click .selectAll a': 'selectAll',
       'click .cutItem a': 'cutClicked',
       'click .copyItem a': 'copyClicked',
       'click .pasteItem a': 'pasteClicked',
@@ -34,32 +36,76 @@ define([
         this.canMove = true;
       }
     },
+    selectAll: function(e){
+      e.preventDefault();
+      var self = this;
+      var page = 1;
+      var count = 0;
+      var getPage = function(){
+        self.app.loading.show();
+        $.ajax({
+          url: self.app.collection.url,
+          type: 'GET',
+          dataType: 'json',
+          data: {
+            query: self.app.collection.queryParser({
+              searchPath: self.model.attributes.path
+            }),
+            batch: JSON.stringify({
+              page: page,
+              size: 100
+            }),
+            attributes: JSON.stringify(self.app.queryHelper.options.attributes)
+          }
+        }).done(function(data){
+          var items = self.app.collection.parse(data, count);
+          count += items.length;
+          _.each(items, function(item){
+            self.app.selectedCollection.add(new Result(item));
+          });
+          page += 1;
+          if(data.total > count){
+            getPage();
+          }else{
+            self.app.loading.hide();
+            self.app.tableView.render();
+          }
+        });
+      };
+      getPage();
+    },
+    doAction: function(buttonName, successMsg, failMsg){
+      var self = this;
+      $.ajax({
+        url: self.app.buttons.get(buttonName).options.url,
+        data: {
+          selection: JSON.stringify([self.model.attributes.UID]),
+          folder: self.model.attributes.path,
+          _authenticator: utils.getAuthenticator()
+        },
+        dataType: 'json',
+        type: 'POST'
+      }).done(function(data){
+        if(data.status === 'success'){
+          self.app.setStatus(_t(successMsg + ' "' + self.model.attributes.Title + '"'));
+          self.app.collection.pager();
+          self.app.updateButtons();
+        }else{
+          self.app.setStatus(_t('Error ' + failMsg + ' "' + self.model.attributes.Title + '"'));
+        }
+      }); 
+    },
     cutClicked: function(e) {
       e.preventDefault();
-      this.cutCopyClicked('cut');
-      this.app.collection.pager(); // reload to be able to now show paste button
+      this.doAction('cut', _t('Cut'), _t('cutting'));
     },
     copyClicked: function(e) {
       e.preventDefault();
-      this.cutCopyClicked('copy');
-      this.app.collection.pager(); // reload to be able to now show paste button
-    },
-    cutCopyClicked: function(operation) {
-      var self = this;
-      self.app.pasteOperation = operation;
-
-      self.app.pasteSelection = new Backbone.Collection();
-      self.app.pasteSelection.add(this.model);
-      self.app.setStatus(operation + ' 1 item');
-      self.app.pasteAllowed = true;
-      self.app.buttons.primary.get('paste').enable();
+      this.doAction('copy', _t('Copied'), _t('copying'));
     },
     pasteClicked: function(e) {
       e.preventDefault();
-      this.app.pasteEvent(this.app.buttons.primary.get('paste'), e, {
-        folder: this.model.attributes.path
-      });
-      this.app.collection.pager(); // reload to be able to now show paste button
+      this.doAction('paste', _t('Pasted into'), _t('Error pasting into'));
     },
     moveTopClicked: function(e) {
       e.preventDefault();