+++ mockup/patterns/structure/js/views/app.js
@@ -8,11 +8,7 @@ define([
   'mockup-ui-url/views/base',
   'mockup-patterns-structure-url/js/views/table',
   'mockup-patterns-structure-url/js/views/selectionwell',
-  'mockup-patterns-structure-url/js/views/tags',
-  'mockup-patterns-structure-url/js/views/properties',
-  'mockup-patterns-structure-url/js/views/workflow',
-  'mockup-patterns-structure-url/js/views/delete',
-  'mockup-patterns-structure-url/js/views/rename',
+  'mockup-patterns-structure-url/js/views/generic-popover',
   'mockup-patterns-structure-url/js/views/rearrange',
   'mockup-patterns-structure-url/js/views/selectionbutton',
   'mockup-patterns-structure-url/js/views/paging',
@@ -24,59 +20,53 @@ define([
   'mockup-patterns-structure-url/js/collections/selected',
   'mockup-utils',
   'translate',
+  'pat-logger',
   'jquery.cookie'
 ], function($, _, Backbone, Toolbar, ButtonGroup, ButtonView, BaseView,
-            TableView, SelectionWellView, TagsView, PropertiesView,
-            WorkflowView, DeleteView, RenameView, RearrangeView, SelectionButtonView,
+            TableView, SelectionWellView,
+            GenericPopover, RearrangeView, SelectionButtonView,
             PagingView, AddMenu, ColumnsView, TextFilterView, UploadView,
-            ResultCollection, SelectedCollection, utils, _t) {
+            ResultCollection, SelectedCollection, utils, _t, logger) {
   'use strict';
 
-  var DISABLE_EVENT = 'DISABLE';
+  var log = logger.getLogger('pat-structure');
 
   var AppView = BaseView.extend({
     tagName: 'div',
-    /* we setup binding here and specifically for every button so there is a
-     * way to override default click event behavior.
-     * Otherwise, if we bound all buttons to the same event, there is no way
-     * to override the event or stop bubbling it. */
-    buttonClickEvents: {
-      'cut': 'cutCopyClickEvent',
-      'copy': 'cutCopyClickEvent',
-      'paste': 'pasteEvent',
-      'tags': DISABLE_EVENT, //disable
-      'properties': DISABLE_EVENT,
-      'workflow': DISABLE_EVENT,
-      'delete': DISABLE_EVENT,
-      'rename': DISABLE_EVENT,
-      'rearrange': DISABLE_EVENT
-    },
-    buttonViewMapping: {
-      'secondary.tags': TagsView,
-      'secondary.properties': PropertiesView,
-      'secondary.workflow': WorkflowView,
-      'primary.delete': DeleteView,
-      'secondary.rename': RenameView
-    },
     status: '',
+    pasteAllowed: !!$.cookie('__cp'),
     statusType: 'warning',
-    pasteOperation: null,
-    'sort_on': 'getObjPositionInParent',
-    'sort_order': 'ascending',
+    sort_on: 'getObjPositionInParent',
+    sort_order: 'ascending',
     additionalCriterias: [],
-    pasteSelection: null,
     cookieSettingPrefix: '_fc_',
-    pasteAllowed: false,
     initialize: function(options) {
       var self = this;
       BaseView.prototype.initialize.apply(self, [options]);
-      self.setAllCookieSettings();
       self.loading = new utils.Loading();
       self.loading.show();
 
+      /* close popovers when clicking away */
+      $(document).click(function(e){
+        var $el = $(e.target);
+        if($el.is('a') || $el.parent().is('a')){
+          return;
+        }
+        var $popover = $('.popover:visible');
+        if($popover.length > 0 && !$.contains($popover[0], e.target)){
+          var popover = $popover.data('component');
+          if(popover){
+            popover.hide();
+          }
+        }
+      });
+
       self.collection = new ResultCollection([], {
         url: self.options.collectionUrl,
-        queryParser: function() {
+        queryParser: function(options) {
+          if(options === undefined){
+            options = {};
+          }
           var term = null;
           if (self.toolbar) {
             term = self.toolbar.get('filter').term;
@@ -86,15 +76,18 @@ define([
             sortOn = 'getObjPositionInParent';
           }
           return JSON.stringify({
-            criteria: self.queryHelper.getCriterias(term, {
+            criteria: self.queryHelper.getCriterias(term, $.extend({}, options, {
               additionalCriterias: self.additionalCriterias
-            }),
-            'sort_on': sortOn,
-            'sort_order': self['sort_order'] // jshint ignore:line
+            })),
+            sort_on: sortOn,
+            sort_order: self['sort_order'] // jshint ignore:line
           });
         },
         queryHelper: self.options.queryHelper
       });
+
+      self.setAllCookieSettings();
+
       self.queryHelper = self.options.queryHelper;
       self.selectedCollection = new SelectedCollection();
       self.tableView = new TableView({app: self});
@@ -110,54 +103,19 @@ define([
         app: self
       });
 
-      self.buttonViews = {};
-      _.map(self.buttonViewMapping, function(ViewClass, key) {
-        var name = key.split('.');
-        var group = name[0];
-        var buttonName = name[1];
-        self.buttonViews[key] = new ViewClass({
-          triggerView: self.buttons[group].get(buttonName),
-          app: self
-        });
-      });
-
       self.toolbar.get('selected').disable();
-      self.buttons.primary.disable();
-      self.buttons.secondary.disable();
+      self.buttons.disable();
 
-      self.selectedCollection.on('add remove reset', function(modal, collection) {
-        if (collection.length) {
-          self.toolbar.get('selected').enable();
-          self.buttons.primary.enable();
-          self.buttons.secondary.enable();
-          if (!self.pasteAllowed) {
-            self.buttons.primary.get('paste').disable();
-          }
-        } else {
-          this.toolbar.get('selected').disable();
-          self.buttons.primary.disable();
-          self.buttons.secondary.disable();
-        }
+      var timeout = 0;
+      self.selectedCollection.on('add remove reset', function(/*modal, collection*/) {
+        /* delay rendering since this can happen in batching */
+        clearTimeout(timeout);
+        timeout = setTimeout(function(){
+          self.updateButtons();
+        }, 100);
       }, self);
 
       self.collection.on('sync', function() {
-        // need to reload models inside selectedCollection so they get any
-        // updated metadata
-        if (self.selectedCollection.models.length > 0) {
-          var uids = self.getSelectedUids(self.selectedCollection);
-          self.queryHelper.search(
-            'UID', 'plone.app.querystring.operation.list.contains',
-            uids,
-            function(data) {
-              _.each(data.results, function(attributes) {
-                var item = self.selectedCollection.getByUID(attributes.UID);
-                item.attributes = attributes;
-              });
-            },
-            false
-          );
-        }
-
         if (self.contextInfoUrl) {
           $.ajax({
             url: self.getAjaxUrl(self.contextInfoUrl),
@@ -168,7 +126,7 @@ define([
             error: function(response) {
               // XXX handle error?
               if (response.status === 404) {
-                console.log('context info url not found');
+                log.info('context info url not found');
               }
             }
           });
@@ -178,6 +136,7 @@ define([
 
       self.collection.on('pager', function() {
         self.loading.show();
+        self.updateButtons();
 
         /* maintain history here */
         if(self.options.urlStructure && window.history && window.history.pushState){
@@ -227,6 +186,23 @@ define([
         });
       }
     },
+    updateButtons: function(){
+      var self = this;
+      if (self.selectedCollection.length) {
+        self.toolbar.get('selected').enable();
+        self.buttons.enable();
+      } else {
+        this.toolbar.get('selected').disable();
+        self.buttons.disable();
+      }
+
+      self.pasteAllowed = !!$.cookie('__cp');
+      if (self.pasteAllowed) {
+        self.buttons.get('paste').enable();
+      }else{
+        self.buttons.get('paste').disable();
+      }
+    },
     inQueryMode: function() {
       if (this.additionalCriterias.length > 0) {
         return true;
@@ -253,7 +229,7 @@ define([
     getAjaxUrl: function(url) {
       return url.replace('{path}', this.options.queryHelper.getCurrentPath());
     },
-    defaultButtonClickEvent: function(button) {
+    buttonClickEvent: function(button) {
       var self = this;
       var data = null, callback = null;
 
@@ -303,6 +279,7 @@ define([
     },
     ajaxSuccessResponse: function(data, callback) {
       var self = this;
+      self.selectedCollection.reset();
       if (data.status === 'success') {
         self.collection.reset();
       }
@@ -322,41 +299,8 @@ define([
         window.alert(_t('there was an error performing action'));
       }
     },
-    pasteEvent: function(button, e, data) {
-      var self = this;
-      if (data === undefined) {
-        data = {};
-      }
-      data = $.extend(true, {}, {
-        selection: JSON.stringify(self.getSelectedUids(self.pasteSelection)),
-        pasteOperation: self.pasteOperation
-      }, data);
-      self.defaultButtonClickEvent(button, data);
-    },
-    cutCopyClickEvent: function(button) {
-      var self = this;
-      var txt;
-      if (button.id === 'cut') {
-        txt = _t('cut ');
-        self.pasteOperation = 'cut';
-      } else {
-        txt = _t('copied ');
-        self.pasteOperation = 'copy';
-      }
-
-      // clone selected items
-      self.pasteSelection = new Backbone.Collection();
-      self.selectedCollection.each(function(item) {
-        self.pasteSelection.add(item);
-      });
-      txt += 'selection';
-      self.setStatus(txt);
-      self.pasteAllowed = true;
-      self.buttons.primary.get('paste').enable();
-    },
     setupButtons: function() {
       var self = this;
-      self.buttons = {};
       var items = [];
 
       var columnsBtn = new ButtonView({
@@ -377,17 +321,11 @@ define([
         collection: this.selectedCollection
       }));
 
-      if (self.options.contextInfoUrl) {
-        // only add menu if set
-        items.push(new AddMenu({
-          contextInfoUrl: self.options.contextInfoUrl,
-          app: self
-        }));
-      }
       if (self.options.rearrange) {
         var rearrangeButton = new ButtonView({
           id: 'rearrange',
           title: 'Rearrange',
+          icon: 'sort-by-attributes',
           tooltip: 'Rearrange folder contents',
           url: self.options.rearrange.url
         });
@@ -397,35 +335,12 @@ define([
         });
         items.push(rearrangeButton);
       }
-
-      _.each(_.pairs(this.options.buttonGroups), function(group) {
-        var buttons = [];
-        _.each(group[1], function(button) {
-          button = new ButtonView(button);
-          buttons.push(button);
-          // bind click events now...
-          var ev = self.buttonClickEvents[button.id];
-          if (ev !== DISABLE_EVENT) {
-            if (ev === undefined) {
-              ev = 'defaultButtonClickEvent'; // default click event
-            }
-            button.on('button:click', self[ev], self);
-          }
-        });
-        self.buttons[group[0]] = new ButtonGroup({
-          items: buttons,
-          id: group[0],
-          app: self
-        });
-        items.push(self.buttons[group[0]]);
-      });
       if (self.options.upload) {
         var uploadButton = new ButtonView({
           id: 'upload',
           title: 'Upload',
           tooltip: 'Upload files',
-          icon: 'upload',
-          context: 'success'
+          icon: 'upload'
         });
         self.uploadView = new UploadView({
           triggerView: uploadButton,
@@ -433,6 +348,39 @@ define([
         });
         items.push(uploadButton);
       }
+      if (self.options.contextInfoUrl) {
+        // only add menu if set
+        items.push(new AddMenu({
+          contextInfoUrl: self.options.contextInfoUrl,
+          app: self
+        }));
+      }
+
+      var buttons = [];
+      _.each(self.options.buttons, function(buttonOptions) {
+        try{
+          var button = new ButtonView(buttonOptions);
+          buttons.push(button);
+
+          if(button.form){
+            buttonOptions.triggerView = button;
+            buttonOptions.app = self;
+            var view = new GenericPopover(buttonOptions);
+            self.$el.append(view.el);
+          }else{
+            button.on('button:click', self.buttonClickEvent, self);
+          }
+        }catch(err){
+          log.error('Error initializing button ' + buttonOptions.title + ' ' + err);
+        }
+      });
+      self.buttons = new ButtonGroup({
+        items: buttons,
+        id: 'mainbuttons',
+        app: self
+      });
+      items.push(self.buttons);
+
       items.push(new TextFilterView({
         id: 'filter',
         app: this
@@ -462,7 +410,7 @@ define([
           }
           self.collection.pager(); // reload it all
         },
-        error: function(data) {
+        error: function() {
           self.setStatus('error moving item');
         }
       });
@@ -487,10 +435,6 @@ define([
       if (self.uploadView) {
         self.$el.append(self.uploadView.render().el);
       }
-
-      _.each(self.buttonViews, function(view) {
-        self.$el.append(view.render().el);
-      });
 
       self.$el.append(self.tableView.render().el);
       self.$el.append(self.pagingView.render().el);
@@ -528,6 +472,11 @@ define([
     },
     setAllCookieSettings: function() {
       this.activeColumns = this.getCookieSetting('activeColumns', this.activeColumns);
+      var perPage = this.getCookieSetting('perPage', 15);
+      if(typeof(perPage) === 'string'){
+        perPage = parseInt(perPage);
+      }
+      this.collection.howManyPer(perPage);
     }
   });